requirements.txt
Flask==3.0.0
anthropic==0.17.0
python-docx==1.1.0
PyMuPDF==1.23.22
python-dotenv==1.0.1
werkzeug==3.0.1
